Fix 4 DRY/compliance issues across frontend and backend
Issue 1 - StatusBadge consolidation:
- Extended Badge.tsx StatusType with 9 new statuses (draft, sent, viewed,
  accepted, rejected, expired, succeeded, failed, refunded) and corresponding
  statusConfig entries with correct variant mappings
- Replaced duplicated STATUS_BADGE_CLASSES objects in 6 pages (QuotesPage,
  QuoteDetail, ProposalsPage, ProposalDetail, PaymentsPage, PaymentDetail)
  with reusable <StatusBadge> component from centralized Badge.tsx

Issue 2 - Pagination consolidation:
- Created new PaginationBar component (Pagination.tsx) with responsive
  mobile/desktop views, ellipsis for large page counts, and proper
  aria-labels for accessibility
- Exported PaginationBar from ui/index.ts barrel file
- Replaced ~50 lines of duplicated inline pagination in 3 list pages
  (QuotesPage, ProposalsPage, PaymentsPage) with <PaginationBar>

Issue 3 - StatusTransitionMixin:
- Added StatusTransitionMixin class to base_service.py providing shared
  mark_sent, mark_accepted, mark_rejected methods with configurable
  valid_send_statuses, valid_accept_statuses, valid_reject_statuses
- Updated QuoteService and ProposalService to inherit from
  StatusTransitionMixin, removing 3 duplicate methods from each service

Issue 4 - Numeric precision for monetary fields:
- Changed Float to Numeric(12,2) for all monetary columns in Payment,
  Price, Quote, and QuoteLineItem models for proper financial precision
- Changed float to Decimal in Pydantic schemas (PaymentBase, PriceCreate,
  PriceResponse, CreateCheckoutRequest, CreatePaymentIntentRequest)
- Quote model: tax_rate uses Numeric(5,4) for rate precision

Net result: -198 lines removed through DRY consolidation across 15 files.

+++ backend/src/payments/schemas.py
@@ -2,6 +2,7 @@
 
 from datetime import datetime
 from typing import Optional, List
+from decimal import Decimal
 from pydantic import BaseModel, ConfigDict
 
 
@@ -64,7 +65,7 @@ class PriceResponse(BaseModel):
     id: int
     product_id: int
     stripe_price_id: Optional[str] = None
-    amount: float
+    amount: Decimal
     currency: str
     recurring_interval: Optional[str] = None
     is_active: bool
@@ -102,7 +103,7 @@ class ProductListResponse(BaseModel):
 
 class PriceCreate(BaseModel):
     product_id: int
-    amount: float
+    amount: Decimal
     currency: str = "USD"
     recurring_interval: Optional[str] = None
     stripe_price_id: Optional[str] = None
@@ -114,7 +115,7 @@ class PriceCreate(BaseModel):
 # =============================================================================
 
 class PaymentBase(BaseModel):
-    amount: float
+    amount: Decimal
     currency: str = "USD"
     customer_id: Optional[int] = None
     opportunity_id: Optional[int] = None
@@ -218,7 +219,7 @@ class SubscriptionListResponse(BaseModel):
 
 class CreateCheckoutRequest(BaseModel):
     quote_id: Optional[int] = None
-    amount: Optional[float] = None
+    amount: Optional[Decimal] = None
     currency: str = "USD"
     success_url: str
     cancel_url: str
@@ -231,7 +232,7 @@ class CreateCheckoutResponse(BaseModel):
 
 
 class CreatePaymentIntentRequest(BaseModel):
-    amount: float
+    amount: Decimal
     currency: str = "USD"
     customer_id: Optional[int] = None
     opportunity_id: Optional[int] = None
